Implements balloon help (in Mac OS 9) and help tags (in Mac OS X).
The balloon help in Mac OS 9 is implemented according to the MacApp scheme. To get
things to look right, I had to "wrap" some of the controls in TViews -- there
may have been a more elegant way to get that working ...

In Mac OS X, I have re-used XPFHelpBehavior as a help behavior for a window --
it installs a help tag callback for the window, and the parses the views to
find out which (if any) view is under the cursor and can offer some help.

I have also implemented an XPFHelpCheckbox which turns balloons or tags on and
off. XPFPrefs keeps track of the preference, and XPFApplication "insulates" the
rest of the system from the changes (i.e. if you turn on Balloon Help in XPF,
it won't be turned on when you switch out).

Partially fixes bug 1983.